+++ cloud/acp-admin-cloud-constant/src/main/kotlin/pers/acp/admin/constant/RoleCode.kt
@@ -1,4 +1,4 @@
-package pers.acp.admin.common.constant
+package pers.acp.admin.constant
 
 /**
  * 角色编码