+++ cloud/oauth-server/src/main/kotlin/pers/acp/admin/oauth/controller/api/LoginController.kt
@@ -10,7 +10,7 @@ import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.*
 import pers.acp.admin.common.base.BaseController
 import pers.acp.admin.oauth.constant.OauthApi
-import pers.acp.admin.common.permission.BaseExpression
+import pers.acp.admin.permission.BaseExpression
 import pers.acp.admin.common.vo.InfoVO
 import pers.acp.admin.oauth.domain.ApplicationDomain
 import pers.acp.admin.oauth.domain.UserDomain