+++ cloud/oauth-server/src/main/kotlin/pers/acp/admin/oauth/controller/api/RoleController.kt
@@ -10,8 +10,8 @@ import org.springframework.security.oauth2.provider.OAuth2Authentication
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.*
 import pers.acp.admin.common.base.BaseController
-import pers.acp.admin.common.constant.RoleCode
-import pers.acp.admin.common.permission.BaseExpression
+import pers.acp.admin.constant.RoleCode
+import pers.acp.admin.permission.BaseExpression
 import pers.acp.admin.oauth.constant.OauthApi
 import pers.acp.admin.oauth.constant.RoleConfigExpression
 import pers.acp.admin.common.vo.InfoVO
@@ -53,9 +53,11 @@ constructor(private val logAdapter: LogAdapter, private val roleDomain: RoleDoma
     fun init() {
         try {
             for (field in RoleCode::class.java.declaredFields) {
-                val code = field.get(RoleCode::class.java).toString()
-                if (RoleCode.prefix != code) {
-                    roleCodeList.add(code)
+                val value = field.get(RoleCode::class.java)
+                if (value is String) {
+                    if (RoleCode.prefix != value) {
+                        roleCodeList.add(value)
+                    }
                 }
             }
         } catch (e: Exception) {