+++ cloud/oauth-server/src/test/java/pers/acp/admin/oauth/test/InitData.java
@@ -5,8 +5,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.test.annotation.Rollback;
 import org.springframework.transaction.annotation.Transactional;
-import pers.acp.admin.common.constant.ModuleFuncCode;
-import pers.acp.admin.common.constant.RoleCode;
+import pers.acp.admin.constant.ModuleFuncCode;
+import pers.acp.admin.constant.RoleCode;
 import pers.acp.admin.oauth.BaseTest;
 import pers.acp.admin.oauth.entity.*;
 import pers.acp.admin.oauth.entity.ModuleFunc;